+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Fixed a bug where CKEditor field values weren’t always getting propagated to other sites (with duplicated nested entries) when creating a new element. ([#479](https://github.com/craftcms/ckeditor/pull/479))
+- Fixed styling issues on CKEditor config edit screens. ([#494](https://github.com/craftcms/ckeditor/pull/494))
 
 ## 4.11.0 - 2025-11-19