+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes for CKEditor for Craft CMS
 
-## Unreleased
+## 4.11.1 - 2026-02-18
 
 - Fixed a bug where CKEditor field values weren’t always getting propagated to other sites (with duplicated nested entries) when creating a new element. ([#479](https://github.com/craftcms/ckeditor/pull/479))
 - Fixed styling issues on CKEditor config edit screens. ([#494](https://github.com/craftcms/ckeditor/pull/494))